% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rblas.R
\name{igamma}
\alias{igamma}
\title{Compute incomplete gamma}
\usage{
igamma(a, x)
}
\arguments{
\item{a}{A numeric vector.}

\item{x}{A numeric vector}
}
\value{
\code{dgami(a, x)}
}
\description{
Evaluate an incomplete gamma function:
\deqn{\gamma(a, x) = \int_{0}^{x} t^{a-1}\exp{-t}dt,}
using SLATEC \code{dgami} in \url{https://netlib.org/slatec/}.
When \eqn{a \gt 0} and \eqn{x \ge 0}, compute the result, otherwise the value
is \code{NaN}.
}
