% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rblas.R
\name{igammac}
\alias{igammac}
\title{Compute complementary incomplete gamma}
\usage{
igammac(a, x)
}
\arguments{
\item{a}{A numeric vector.}

\item{x}{A numeric vector}
}
\value{
\code{dgamic(a, x)}
}
\description{
Evaluate an complementary incomplete gamma function:
\deqn{\gamma^{*}(a, x) = \int_{x}^{\inf} t^{a-1}\exp{-t}dt,}
using SLATEC \code{dgamic} in \url{https://netlib.org/slatec/}.
When \eqn{x > 0} or \eqn{x \eq 0 & a \ge 0}, compute the result,
otherwise the value is \code{NaN}.
}
