% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtest1.R
\docType{package}
\name{rtest1}
\alias{rtest1}
\title{rtest1: rpackage test}
\description{
The mypackage package provides three categories of important functions:
foo, bar and baz.
}
\section{Mypackage functions}{

The mypackage functions ...
}

