% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rblas.R
\name{igamma}
\alias{igamma}
\title{Compute incomplete gamma}
\usage{
igamma(a, x)
}
\arguments{
\item{a}{A numeric vector.}

\item{x}{A numeric vector}
}
\value{
\code{dgami(a, x)}
}
\description{
Compute incomplete gamma
}
