% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rblas.R
\name{igammac}
\alias{igammac}
\title{Compute complementary incomplete gamma}
\usage{
igammac(a, x)
}
\arguments{
\item{a}{A numeric vector.}

\item{x}{A numeric vector}
}
\value{
\code{dgamic(a, x)}
}
\description{
Compute complementary incomplete gamma
}
